fix: throttle balance syncer + remove WC patch (#3848)
